test: 1から100までの数をプリントする
